requirements.txt
numpy
pandas
scikit-learn
matplotlib
torch
pyyaml